requirements.txt
# Airflow providers (base image에 core 포함)
apache-airflow-providers-standard>=1.0.0
apache-airflow-providers-postgres>=5.0.0
apache-airflow-providers-amazon>=9.0.0
apache-airflow-providers-http>=4.0.0
apache-airflow-providers-fab>=2.0.0

# Pipeline
dlt[postgres,filesystem]>=1.0.0
dbt-core>=1.8.0
dbt-duckdb>=1.8.0
dbt-postgres>=1.8.0

# Fetcher / Parser
httpx>=0.27.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
trafilatura>=2.0.0

# Storage / DB
boto3>=1.34.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Shared
python-dotenv>=1.0.0

# Dev / Test
pytest>=8.0.0
ruff>=0.4.0
black>=24.0.0
